Generate a synthetic 'variant name' for multiple configurations
Without this change, the `dependencyInsight` report will choose a single
selected configuration to report as the name of the variant. This is
misleading, since multiple configurations can be selected for the
same module version.

With this change, a synthetic name is generated in the case of multiple
selected configurations for a module version. The pattern is simple:
the set of names concatenated together with '+'.